OAK-9046 Index function string-length should index size for binary properties

+++ oak-search/src/main/java/org/apache/jackrabbit/oak/plugins/index/search/util/FunctionIndexProcessor.java
@@ -37,10 +37,10 @@ import org.slf4j.LoggerFactory;
  * definition (XPath) to the internal Polish notation.
  */
 public class FunctionIndexProcessor {
-    
+
     private static final Logger LOG =
             LoggerFactory.getLogger(FunctionIndexProcessor.class);
-    
+
     private String remaining;
 
     private static final PropertyState EMPTY_PROPERTY_STATE = EmptyPropertyState.emptyProperty("empty", Type.STRINGS);
@@ -48,10 +48,10 @@ public class FunctionIndexProcessor {
     protected FunctionIndexProcessor(String function) {
         this.remaining = function;
     }
-    
+
     /**
      * Get the list of properties used in the given function code.
-     * 
+     *
      * @param functionCode the tokens, for example ["function", "lower", "@name"]
      * @return the list of properties, for example ["name"]
      */
@@ -68,7 +68,7 @@ public class FunctionIndexProcessor {
 
     /**
      * Try to calculate the value for the given function code.
-     * 
+     *
      * @param path the path of the node
      * @param state the node state
      * @param functionCode the tokens, for example ["function", "lower", "@name"]
@@ -94,10 +94,10 @@ public class FunctionIndexProcessor {
         PropertyState ret = stack.pop();
         return ret==EMPTY_PROPERTY_STATE ? null : ret;
     }
-    
+
     /**
      * Split the polish notation into a tokens that can more easily be processed.
-     *  
+     *
      *  @param functionDescription in polish notation, for example "function*lower*{@literal @}name"
      *  @return tokens, for example ["function", "lower", "{@literal @}name"]
      */
@@ -107,7 +107,7 @@ public class FunctionIndexProcessor {
         }
         return functionDescription.split("\\*");
     }
-    
+
     private static PropertyState calculateFunction(String functionName,
                                                    Deque<PropertyState> stack) {
         PropertyState ps = stack.pop();
@@ -139,7 +139,7 @@ public class FunctionIndexProcessor {
                 x = ps.getValue(Type.STRING, i);
                 type = Type.STRING;
             } else if ("length".equals(functionName)) {
-                x = (long) ps.size(i);
+                x = ps.size(i);
                 type = Type.LONG;
             } else {
                 LOG.debug("Unknown function {}", functionName);
@@ -156,7 +156,7 @@ public class FunctionIndexProcessor {
         }
         return result;
     }
-    
+
     private static PropertyState getProperty(String path, NodeState state,
                                              String propertyName) {
         if (PathUtils.getDepth(propertyName) != 1) {
@@ -183,16 +183,16 @@ public class FunctionIndexProcessor {
         }
         return ps;
     }
-    
+
     private static String getLocalName(String name) {
         int colon = name.indexOf(':');
         // TODO LOCALNAME: evaluation of local name might not be correct
         return colon < 0 ? name : name.substring(colon + 1);
     }
-    
+
     /**
      * Convert a function (in human-readable form) to the polish notation.
-     * 
+     *
      * @param function the function, for example "lower([name])"
      * @return the polish notation, for example "function*lower*{@literal @}name"
      */
@@ -203,7 +203,7 @@ public class FunctionIndexProcessor {
         FunctionIndexProcessor p = new FunctionIndexProcessor(function);
         return QueryConstants.FUNCTION_RESTRICTION_PREFIX + p.parse();
     }
-    
+
     String parse() {
         if (match("fn:local-name()") || match("localname()")) {
             return "@:localname";
@@ -249,11 +249,11 @@ public class FunctionIndexProcessor {
             return property(prop.replaceAll("@", ""));
         }
     }
-    
+
     String property(String p) {
         return "@" + p;
     }
-    
+
     private String read(String string) {
         match(string);
         return "";
@@ -268,7 +268,7 @@ public class FunctionIndexProcessor {
         }
         return "*";
     }
-    
+
     private boolean match(String string) {
         if (remaining.startsWith(string)) {
             remaining = remaining.substring(string.length());
